+++ test/unit/fuzz-tools.test.js
@@ -227,6 +227,7 @@ import * as dermscoreV234 from '../../lib/dermscore-v234.js';
 import * as painscoreV235 from '../../lib/painscore-v235.js';
 import * as ophthoV236 from '../../lib/ophtho-v236.js';
 import * as cardioechoV237 from '../../lib/cardioecho-v237.js';
+import * as anthroV238 from '../../lib/anthro-v238.js';
 
 const MODULES = {
   'clinical.js': clinical,
@@ -398,6 +399,7 @@ const MODULES = {
   'painscore-v235.js': painscoreV235,
   'ophtho-v236.js': ophthoV236,
   'cardioecho-v237.js': cardioechoV237,
+  'anthro-v238.js': anthroV238,
 };
 
 const MATRIX = [0, -1, 1e9, NaN, Infinity, -Infinity, '', undefined, null];